scripts/requirements.txt
# Core DL (GPU)
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

# Federated Learning
flwr[simulation]==1.8.0

# Data handling
pandas>=2.2.0
pyarrow>=15.0.0
numpy>=1.26.0

# Image processing
Pillow>=10.0.0

# Utils
tqdm>=4.66.0
scikit-learn>=1.4.0

# Config
omegaconf>=2.3.0

# Optional (later visualization)
matplotlib>=3.8.0
seaborn>=0.13.0

#transformers

#pip install torch==2.6.0 torchvision==0.21.0 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu126